Fixed compilation errors due to enum/int comparisons and uninitialized variable.

+++ src/db.c
@@ -1424,13 +1424,13 @@ static int update_connection_record_params(mapper_db_connection con,
 
     mapper_boundary_action bound;
     bound = mapper_msg_get_boundary_action(params, AT_BOUND_MAX);
-    if (bound != -1 && bound != con->bound_max) {
+    if ((int)bound != -1 && bound != con->bound_max) {
         con->bound_max = bound;
         updated++;
     }
 
     bound = mapper_msg_get_boundary_action(params, AT_BOUND_MIN);
-    if (bound != -1 && bound != con->bound_min) {
+    if ((int)bound != -1 && bound != con->bound_min) {
         con->bound_min = bound;
         updated++;
     }
@@ -1525,7 +1525,7 @@ static int update_connection_record_params(mapper_db_connection con,
     updated += update_int_if_arg(&con->dest_length, params, AT_DEST_LENGTH);
 
     mapper_mode_type mode = mapper_msg_get_mode(params);
-    if (mode != -1 && mode != con->mode) {
+    if ((int)mode != -1 && mode != con->mode) {
         con->mode = mode;
         updated++;
     }
